refactor(cli): split the v8 entry into runtime and cli modules
main.ts was a grab-bag under an entry name: the runtime wiring
(assembleRuntime, detectPackageManager, makeOnSignal,
makeGetCredentials) moves to v8/runtime.ts, tree construction
(buildCli, now using defineCommandFamily) to v8/cli.ts, and main.ts
keeps only the thin main(process) entry composing them. Tests follow
the split and the engine's ./testing subpath.
